Disable tools via --withouth-tools configure switch
lvm-cache-stats is a simple tool for displaying stats for LVM cache
devices which was introduced in version 2.21. It is the only tool at
this point and we don't really need it so disable it to avoid the need
for a new binary package. Should more tools be added which appear to be
useful, this decision can be revisited.